Wallet is now automatically created
To provide a smoother onboarding experience, the seed is created/written
on disk a soon as the user starts a new wallet instead of being shown to
the user first. The node then automatically starts. The user will then
have the ability to check the seed and backup the mnemonics.

This commit also removes the dedicated activity for wallet
initialization, and uses fragment instead which improves the pages flow.