Store closing channel events as OutgoingPayment
This is a dirty trick that allows the wallet to display channels closing
event in the payment list. We use the arbitrary `external_id` field to
mark the payment as a special case.

This fake payment uses a random preimage with a corresponding payment
hash. Target node id is random.

This commit also removes the channels map in the node supervisor actor
that was mirroring the current active channels list. As a side effect,
the balance is now retrieved by querying the database via the kit API.