Show blocking message when connection to electrum is lost
Generic NodeData object has been removed. Balance is not stored in a
dedicated LiveData attribute. Added a NetworkInfo object in AppKit
holding data regarding network connection and electrum connection.

Snackbar for network connectivity has been replaced by a custom layout
because it's more flexible.

This commit is a work in progress: the message should not be
displayed when the app is starting.