Fix issue where app would not reconnect to remote peer
Some devices disconnect the app as soon as it leaves foreground and
moves to background in order to save battery and data. This would
prevent Phoenix from receiving from another app because it would be
disconnected and not properly reconnect. A change was added to force
the app to reconnect to its remote peer as soon as it comes back to
foreground, however this is not enough: we should also wait for
network connection to come back.

Potentially fixes #6