Handle countdown in payToOpen inside view model
Timer will now survive view configuration change. Countdown display
should also be handled safely with a view model to prevent updating
the view when the dialog is detached.

Note that the countdown is still relative to the time where the dialog
is open and will not accurately represent the server internal countdown
which is private for now.