Show confirming channels as incoming funds
A channel can stay in the WAIT_FOR_FUNDING_CONFIRMED longer than
expected, in which case the user will see a 0 balance although the
payment has been received.

This can happen if the Electrum server the user is connected to does
not see the funding transaction of the channel. This can happen if the
server is desynced, or if there is an issue on the server. In that case
the app should inform the user that funds are coming his way, but are
not yet confirmed.

Note that in a normal scenario, the channel is active even the funding
tx is not confirmed (but still visible), so the balance will be correct.