Improve state model in LNURLWithdraw
Use sealed class to store the fragment state instead of an enum.
Being more flexible we can store detailed data in the view model state.
The UI can then be safely updated from the fragment itself which is
lifecycle aware.

This fixes a bug where displaying an error/success message from the
remote callback could lead to crashes if the lightning payment is
received before the withdrawal service API responds.