Use progressive fees when sending trampoline payment
Trampoline fees is not a static (0.1% + base) value anymore.
Instead the core can do several attempts, from the cheapest option
to the more expensive ones.

The app must now feed the core with a list of possible fee settings
that the user will be able to configure (not done yet).

Note that the app must also compute the fees that would incur from
using the routes from the invoice's routing hints, and include
those fees to the total fees. Same thing for the CLTV expiry.

Note that the 'send all' option will use the most expensive fee
setting available to increase the chances that the payment will go
through.