Improve state management in payment details
Instead of storing the page's state as an enum, forcing the model to keep
track of the actual details in another field, we use a sealed class which
also stores the details of the payments, thus preventing inconsistencies
between the details of the payment and the view state.

This commit also fixes issues with outgoing payments: the changes with
trampoline attempts makes it possible to have both failed and successful
outgoing payments sharing the same parent id. Fees have also changed and
must be computed from the list of (successful) outgoing payment parts.

We also revert the changes where settings navigation paths were moved into
a subgraph.