Improve TOR support (esp. Android 10)
Removed the tor-android-binary lib due to some issues, instead we add
only the classes we need and modify them a bit. Plan is to properly
rewrite these classes.

This notably fixes an issue in Android 10 (Q) where this library would
copy the tor binary and rewrite it into the app data folder, which is
now forbidden by Android 10 because is enforces the W^X rule

https://android-review.googlesource.com/c/platform/system/sepolicy/+/804149

The packaged tor libs are also added to the project.

This commit also add a custom jctl event handler for the tor onion
manager. The startup page display a specific animation when tor is
starting.