Add Tor support (opt-in) (#35)
User can enable TOR with a button in the settings, and the app will use 
TOR to connect to its Lightning peer/Electrum server. Some API calls 
(for example the API to retrieve the bitcoin/fiat exchange rate) will not 
use TOR. By default TOR is disabled.

We use the Tor_Onion_Proxy_Library with jtorctl to control the TOR binary.
Some classes from the tor-android-binary have been added in the sources.
We do not use this library as-is, because it copies the binary to the same
directory as the config resources, which is for the app data folder. This 
folder being writable, it makes the binary non executable in Android Q
since it enforces the W^X rule.

https://android-review.googlesource.com/c/platform/system/sepolicy/+/804149

The packaged tor libs (v0.4.2.5) are added to the project with the following 
checksum:

- arm64-v8a/libtor.so
9129f5184ac5b7da51345ca820c3bacd5fbe0e3586179028b186a4ea07de7e9a

- armeabi-v7a/libtor.so
9a33eeaeddd388f069bf39dd60c676e6016c39cbe1922a3fc19dd768a1814964

- x86/libtor.so
ecdf6d50989d0eb80aab7473430c06ca06df8aaefb9a0ac9e9761079a4df7129

- x86_64/libtor.so
4c4acd116985762e21438f2bd3168f506c7a5a250fd32e2619eda525b4b9d97c

This commit also adds a custom jtorctl event handler for the tor onion
manager. The startup page will display a specific animation when tor is
starting.

This commit also improves connectivity feedback using a dedicated dialog
and removes the full width connectivity banner at the bottom of the app,
replaced it with a more discreet notification at the top of the main
fragment page.

The user can also enable TOR or use a custom electrum server even
before creating the wallet.

Fixes #5