Fix race condition in state handler in startup
The startup state handler, tasked with controlling the navigation
between fragments when the node starts, can be started by the onStart
and the state observer, which can lead to race conditions where the
state and the current navigation are not synced, ultimately leading to a
error in the android navigation controller.

A longer term fix would be to prevent starting this handler from the
fragment lifecycle, and relying only on the observer. However this would
break the pin/biometrics dialog on alt-tab.