Remove default pin code
Starting with version 11, the default pin code is removed because it
brings no security benefits and makes the pin management logic
inconsistent, and also slows down the application startup since
decryption is inherently slow.

Instead the app uses a default `null` pin. A migration "script" has been
added for users that are already have the application installed but have
not enrolled a pin to protect access to the wallet.

This commit prepares the separation between seed encryption and the
pin/biometric system protecting access to the wallet.