Move node logic to a Service
This service now contains the kit and the state of the node (off,
starting, started...).

This service can be `started`, meaning that it has been started from
the background and is headless, so a notification is displayed ; or it
can be `bound`, when the app UI is active. In the latter case,
interaction with the node is done through the shared `AppViewModel`,
using the `service` mutable.

Note that the state of the app is set in the service, and can be
observed from the UI using the `AppViewModel` shared model. A special
`MediatorLiveData` has been created to reconcile the state internal
to the service with the state from the model, accounting for the fact
that the service may not always be reachable. Otherwise this state in
the UI view model simply mirrors the service state.

Some live data such as the state of the network have been moved to the
Application object because they are be changed from the service, but
only used from the UI.