Improve seed protection using keystore+Biometrics
Refactored the EncryptedSeed object, making it a sealed trait with a
(obsolete) V1 version using a PIN, and a new V2 version encrypting the
seed using the Android Keystore. The V2 key can also require the user
to authenticate to be usable (opt-in) which has the drawback of
disabling the "offline" receive feature.

Most of the PIN code logic has been removed ; some of it has been kept
to allow for the migration of existing wallets. The access protection
screen has been reworked, the layout is now much simpler and most of the
controller previous logic has been removed. The migration from V1 to V2
is made at startup when the user first starts the app with the new
version.

The encrypted seed is still stored in a private file on device, with the
version byte being incremented. A SeedManager class has been added to
abstract the operations on the seed file.