Use Github action to build Testnet APK on master (#94)
Adds a Github action workflow to build the app from the `master` branch. 
The app runs on Testnet. The apk is uploaded to the Github actions artifact 
repository, and to AWS to have a one-click APK download button.

The README been updated to add links to the build/apk.

Fixes #93