Fix issue with StrongBox in keystore. Update wording in access screen
Storing a key in the device strong box is not supported on every
devices. Try to use it if possible but do not crash the key creation
otherwise.

The wording for the full lock option is more explicit on the fact that
a key with require user auth prevents the app from running in
background.

This commit also fixes a few issues with the startup screen error state
handling.