Improve authentication system and startup lock state
Auth errors are now more explicit. Hard auth requires fingerprint and
confirmation by the user that the seed has been backed up, to avoid
lockout scenarii.

The Android sdk has been upgraded to 30, along with the androidX
biometrics library. We now use the `Authenticators` object to set the
accepted credential types for soft auth and hard auth.

Authentication errors are removed from the KitState and now have their
own state object. This avoids modifying the KitState from the UI, and
allows the application to be locked while the node is running.