Handle wakeup reason in fcm service and improve notif
Wakeup may not be a incoming payment but also a htlc fulfillm so the
message in the notification should be different.

The icon for the Phoenix notifications has been updated using a thicker
and larger icon.

A notification is displayed when the FCM service ignores a wakeup
request on account of the seed encryption type.