Support background payment reception (#104)
Support has been added for background payment reception, using Firebase
Cloud Messaging to wake up the app. The wallet register a FCM token with 
the peer and can wake up when the peer wants to send a fulfilled HTCL or an
incoming payment.

A service has been added, handling all interactions with the underlying nodes.
This service can be headless, in which case a foreground notification is 
displayed ; or it can be bound to the UI, in which case interaction with the 
node from the UI is done through the shared `AppViewModel`, using the 
`service` mutable.

The state of the kit is now set in the service, and can be observed from the shared
`AppViewModel`. When the service is not reachable a special state is used.

A new `EncryptedSeed.V2` object has been added, using the Android Keystore 
and the biometrics api to securely store the seed while allowing background 
processing. `WithAuth` requires user authentication to use the key and as such
the background processing features.

Most of the `EncryptedSeed.V1` code and PIN logic has been removed ; some 
of it has been kept for migration purposes. Migration from V1 to V2 is done
when users unlocks the app for the first time after update.

A `SeedManager` class has been added to abstract operations on the seed file.

The access setting screen has been reworked and is more simple. Authentication
errors at startup are now separated from the kit state, due to the fact that the app
can be started, but not unlocked yet.

First time user experience screens have been added to introduce the app to 
new users. 

Pay-to-open requests are now automatically accepted by default. Users can disable
this, in which case a system notification is displayed when a request is incoming.

A metadata database has been added, storing data about payments that are 
outside the scope of the underlying node db and should not be stored there. 
This include data about opening or closing channels, and data for swap-in and
swap-outs transactions.

Users can now set the default invoice description, which now defaults to an
empty string.

Invoice expiration has been increased to 1 week.

A patch note screen has been added to list notable changes after an update. This
feature is used for version 15, including mention of the auto pay-to-open feature. 

Onchain explorer can now be changed. Default explorer is now mempool.space.