Use eTLD+1, with fallback to host when generating LNURL signing key
The service's url for handling LNURL queries is used to generate the key
and link it to the service. However the full url, including subdomains
is a technical parameter that can change without notice, while the
eTLD+1 should be static, and as such is used when possible to ensure a
stable experience.