Rework home screen top motion and add option to hide balance
Balance can now be hidden in the home screen to increase privacy when
the app is open and in foreground.

The home top section animation has been slightly reworked. The payments
list can now take the whole screen when users scrolls down. When users
scrolls up, depending on the origin of the dragging motion, either the
top section is expanded, or the list is scrolled upward.