Added setting to change max fee for trampoline payments
A new setting in the payment option screen has been added, with a dialog
to override the default maximum trampoline fee setting. The method
sending trampoline payments now uses this custom fee when needed.
Default fee settings over this custom fee are ignored.

Also, the fee in the invoice's routing hints is now ignored. The
trampoline fee settings is the definitive maximum fee that the wallet
can use.

The trampoline proportional fee is now defined as satoshi per millionths
and we can reuse the fr.acinq.eclair.nodeFee method. For convenience,
this proportional fee is displayed to the user as a percentage and
converted back to sat-per-millionths for internal usage.

The previous (unused) screen for fee settings has been removed.