Store channels in a sqlite database (#62)
Using a sqlite DB enables sharing SQL schema and data with the scala
implementation.  It will also make migration significantly easier for 
existing Phoenix wallets.

WE use the SQLDelight library is used to generate type safe kotlin 
methods from sql statements and schemas. A `ChannelsDatabase.sq` 
file is created to describe the database.

Note that the channels database is a physical file. On Android this file
will be stored in the app's private directory > databases folder.