Implement SQLite payments database (#80)
Using a sqlite DB enables sharing SQL schema and actual data with the scala
implementation. The payment database is an actual physical file that should be
stored in a private directory of the app.

The SQLDelight library is used to generate type safe kotlin methods from
sql statements and schemas. Three `.sq` file have been created, generating 
tables and queries for outgoing payments, incoming payments, and aggregated 
list of payments.

Outgoing payments are stored using two tables: one for general information 
about the payment (such as the description or the recipient) and a table for the
parts of the payment (a payment can be settled with several part thanks to MPP).

Only minimal information for failures of outgoing parts are stored.