Push notifications on iOS (#89)
iOS app now supports push notifications, and can receive payments while
the app is running in the background.

The notification system uses Firebase Cloud Messaging. The app registers
a token with the peer, and the server can use that token to push a 
notification to the app to wake it up, and connect to the peer.

A notification (and badge) is displayed when the app has received a payment 
when in the background.

The user is prompted for push permission when they enter the Receive screen.
And warn the user if they've disabled something that will prevent Phoenix from 
either receiving payments in the background, or being notified about receive 
payments.

Within the iOS layer, we automatically disconnect from the network(s) when
going into the background. This allows the server to send us push notifications. 
And iOS then allows for re-connection when processing an incoming push notification.