Improve error management in LNURL requests
Errors returned by LNURL services were not properly handled and
information was lost, causing confusion.

This commit correctly forwards the error message returned by the LNURL
service, especially for the withdrawals. If the service does not
return a meaningful error, we show the http code, or a generic message
at worst.

This also applies for the "metadata" request that is sent right after
scanning a LNURL.