Add support for swap-in payments (#131)
Implement query simultaneously adding and receiving incoming payments.

Swap-in intent/model has been added in receive controller.

The home controller now exposes the incoming balance. It will hold
the balance for the pending swap-ins (useful because of the delay for
swaps). When a swap-in confirms on that address, the incoming balance is
updated.

The logic updating the payments in the payment manager has been 
simplified: instead of listening to payment events from the lightning engine, 
we directly watch the database.