Set phoenix-shared minimum ios version to 14.0
By default, the phoenix shared framework uses a 9.0 minimum iOS version
which the app stores considers as invalid. The root cause is not completely
clear, but it may be a side effect of the PhoenixCrypto module requiring
iOS 13.0 to run. It seems that the app store does not take into account the
target iOS version set by the phoenix-ios project.