Add an option to disable automatic on-the-fly channel
Since pay-to-open incurs a fee, users may want to disable it. An option
to do so has thus been added to the payments settings screen.

Since this can prevent the wallet from receiving any payments, a
notification is displayed to inform the user why the payment is failing.
This notification has a 20 min cooldown (and can be disabled in the
Android settings for Phoenix).