Load images stored on device and extract QR code
Lightning invoices are difficult to share (long unwieldy strings)
especially on social networks, so it's easier to share the QR code
image instead. The wallet is now able to load local image and extract
any valid invoices that they contain.

Phoenix uses the device file explorer to browse for the image so this
feature does not require new permissions from the user.