(legacy) remove full lock control
This option is dangerous because when enabled, users who have not
backed up their seed (or have lost their backup) and modify their
fingerprint in Android will lose access to the wallet. This is not
obvious and warnings do not really help.

This option remains active for people who have already enabled it.