(legacy) Prevent sending more than twice the requested amount
For compliance with bolt#11, the wallet should not let the user try to
send more than twice the requested amount. Otherwise the payment will
most probably be rejected by the receiver anyway, and the error returned
by the receiver will be confusing. So UX wise it's better to error early.

See #232