(android) Improve legacy switch logic
Legacy preferences are migrated to a datastore. Datastore data are now
separated between user preferences and internal application data.

The legacy app now exposes a status in the AppContext class that can be
used in the modern/legacy app switch logic.

The initialization screens (create/restore) have been moved to separate
files, with a new viewmodel that tracks the writing status of the seed.