(android) Detect legacy channels using a TLV in init message
Thanks to the TLV in the Init message we can request the peer to signal
legacy channels, and switch to the legacy app if needed. This event is
watched from the main activity, and updates a 'switch to legacy' status.

The 'switch to legacy' status is stored in a datastore pref so it can be
directly observed from a composition.