Support closing txs parts in cloud data wrapper
This commit adds a new cloud data wrapper for closing transaction parts.
Those parts are stored in `OutgoingPaymentWrapper` a separate field from
the existing `parts` one, to ensure backward compatibility with
previously encoded data: we could have used a sealed class abstracting
parts, but that would have meant that an additional `type` field was now
required in the concrete object. Data serialized beforehand and missing
this type would not be readable anymore.