Make payment details a single endpoint and rework technical screen
The payment details screen is separated in several subscreen, including
a "splash" screen that display only the more important information, and
a detailed technical screen that displays most of the information about
the payment.

The Card component has been updated and can now specify the spacing
between every one of its child composables.

The Amount pretty formatter can now specify a policy about how the
millisatoshis are displayed: never (default policy), always, or
depending on the amount. This follows the rules established in the
issue #245.