Add payment option screen in settings (#275)
The maximum acceptable values for trampoline fees are defined in the 
`PaymentOptionsConstants` in the shared module.

Several components have been added or modified for a more consistent
look. Inputs now use the OutlinedTextField design and have some validation
logic.

The amount and description of the current payment request have been 
added to the Receive screen.