Store closing transactions as parts (#271)
Payments representing channels closing are now stored in their own SQL
table, `outgoing_payment_closing_tx_parts`. Several methods in the
outgoing payments database have been renamed.

A v5->v6 migration file for the new closing txs table has been added.

The legacy outgoing onchain status that used to contain the list of the closing
transactions is now mapped to the new closing parts. This is also done for the
cloud data.

Note that for the cloud data,  in `OutgoingPaymentWrapper`, the closing parts 
are stored in a separate field and not in the already existing `parts`, to ensure 
backward compatibility with previously encoded data. A sealed class abstracting
parts could have been added, but that would have meant that an additional `type` 
field was now required in the concrete object to let the deserializer discriminate
between the different parts type.

The iOS app UI has been updated. The Android UI has not been updated yet (this
is done in a separate PR).

Co-authored-by: Robbie Hanson <robbiehanson@deusty.com>