Prepare migration of Android app and improve several screens (#252)
The legacy app is now a library that is embedded in phoenix-android and
can be launched from the new app. The modern phoenix-android app checks
at runtime (through a custom TLV in the `Init` message) whether we should load 
the legacy app or not, and only when it makes sense. A status object has been
added to monitor the state of the app and help communication between the
legacy app and the modern app.

A screen has been added in the legacy app to initiate the migration. This process
is started by the user. 

A script migrating the legacy payments database data into the new payments
database has been added. This part is still a work in progress.

Legacy preferences are also migrated to the new datastore library. Datastore 
data are now separated between user preferences and internal application data.

Several screens in the new app have been refactored and improved. A screen
for the technical details of a payment has been added, as well as a new screen in
the settings to manage the payment options. Several components (especially the 
inputs) have been improved.