Support swap-out in iOS and Android (new version) (#286)
The SwapOut details object now contains a payment request and the
swap-out fee, instead of just a payment hash. There is no database 
migration needed as no swap-out have ever been done in production 
with the new codebase.

The scan controller has been refactored in some places. Models and
intents for swap-outs have been added

A new UI has been added to the iOS app and the Android app for swap-
out payments, with roughly the same look and feel.

When a bitcoin URI contains a lightning payment request parameter, 
the user choose how the payment is done, on-chain or off-chain.

Co-authored-by: Robbie Hanson <robbiehanson@deusty.com>