(wip) Enable new memory model and use coroutine-1.6.3
The old memory model based on freezing objects that moved between threads
caused a lot of issues on iOS. With the upgrade to kotlin 1.6, ktor 2.x
and coroutines 1.6 it makes sense to also move to the new model which is
much more workable than the old one.

To do so we enable an experimental flag and use the adequate kotlinx
coroutine library version.

See
https://github.com/JetBrains/kotlin/blob/master/kotlin-native/NEW_MM.md

This commit is a work in progress, because several changes are needed
for the iOS app to be fully functional. Several thread checks will have
to be removed, for example in SyncTxManager. However a lot of the app
works right away.