(android) Add UI for lnurl-auth with support for legacy scheme
The legacy android wallet uses a non-standard logic for the key signing
the authentication token. Instead of using the master key, the node's
key is used.

As such the UI lets users select what scheme to use for backward
compatibility. The default scheme is the protocol one (and the iOS one).