Support for new swap-in events
The swap-in address now belongs to the wallet. It is generated
by lightning-kmp using an address on account 1.

The swap-in flow has also changed. Swap-ins can be accepted or
rejected by Phoenix or the peer (for example if the fee is too
low or the swapped amount below a given threshold). There are
new events for channel creation (that are triggered by swap-ins
but also by regular pay-to-opens).