Update to Kotlin 1.6.21 with new memory model (#299)
Several dependencies are also upgraded, including:
- secp to `0.7.0`
- lightning-kmp to `1.5+`
- ktor to `2.0.3`
- kotlinx-coroutines to `1.6.3`
- kotlinx-datetime to `0.3.2` (transitive from lightning-kmp)
- kotlinx-serialization to `1.3.3`

== New memory model ==

We also enable the new memory model which lets us work seamlessly
between threads. The old memory model based on freezing objects that 
moved between threads caused a lot of issues on iOS. 

To do so we enable an experimental flag and use the adequate kotlinx
coroutine library version (not native-mt).

Some workarounds can be removed in the iOS app (e.g. in SyncTxManager). 
The `preventFreeze()` and `makeShared()` methods are also removed.

For more information, see:
https://github.com/JetBrains/kotlin/blob/master/kotlin-native/NEW_MM.md

== Dual funding ==

This commit also adds basic support for dual funding, both with a feature 
in the node parameters, and in the way we handle the swap. 

The swap-in address now belongs to the wallet and is generated by 
lightning-kmp using an address on account 1.

The flow is different: swap-ins can be accepted or rejected by Phoenix or the 
peer (for example if the fee is too low or the swapped amount below a given 
threshold). There are new events for channel creation (that are triggered by 
swap-ins but also by regular pay-to-opens).

Co-authored-by: Robbie Hanson <robbiehanson@deusty.com>