Use full domain for lnurl-auth linking key
The login service domain is used to compute the signing key. Instead of
using the etld+1, we now use the full domain name. Any subdomain change
by the service will be dealt by the service. Backward compatibility is
ensured for services known to already use lnurl-auth, so that users can
login to their existing account. New services will use the full domain.

This commit affects the legacy android app and the ios app.

This commit is a hotfix taken from the `lnurl-auth-full-domain` branch.