(android) Update migration logic and fix a few issues
The Legacy Migration signal is not used anymore and can be
removed. The swap-in address is also not provided by the
peer anymore, but is computed locally using the new app
derivation path, from the user's wallet. The swap channel
opened for the migration is dual-funded.

A few issues in the UI have also been fixed.