Add Tor support to iOS and (new) Android (#306)
The tor-mobile-kmp library embeds the Tor binary and manages the 
connection to the Tor proxy.

The connection daemon has been updated to take the state of the 
Tor connection into account, blocking any connection attempt if Tor 
is not ready. Both apps now define a preference for enabling Tor that
is injected when starting `PhoenixBusiness` and the connection 
daemon. By default, Tor is disabled.

Note that only the Peer connections and the Electrum connections 
use the Tor proxy. Standard HTTP connections do not.

Co-authored-by: Robbie Hanson <robbiehanson@deusty.com>