Expose a smart channels flow in PeerManager
This 'smart' flow combine the bootFlow and the channelsFlow
from the Peer object so that the initial list of channels now
comes from the local database and does not remain empty while
the wallet is starting up and busy reestablishing connection
with the Peer.

The objects in the flow also contain a few helpers methods
like for computing the usable balance. This will provide the
UI a single, consistent source of information.

The BalanceManager and the Channel configuration controller
now use that flow instead of directly interacting with the
flows from the Peer.