Add single & consistent source of channels in PeerManager (#338)
A new channelsFlow had been added to the PeerManager.
This 'smart' flow combines the bootFlow and the channelsFlow
from the Peer object so that this flow is initialised with a list of 
channels from the local database, and then, once the connection
is established with the peer, will contain an up-to-date list of
channels maintained in cooperation with the peer. This avoids
having an empty list of channels when the app is starting.

The flow exposes a new `LocalChannelInfo` object that contains
a few helpers methods, like for computing the usable balance.
This provides a single, consistent source of information.

The BalanceManager and the Channel configuration controller
now use that flow instead of directly interacting with the
flows from the Peer.