Relax lnurl tag validation
We should not be too strict regarding the tag parameter
in the initial lnurl. It may be empty or unknown, in which
case the tag is ignored and the lnurl is executed.

Note that this does not affect the subsequent validation
where the json returned by the service must contain a
valid tag attribute.