Add simple CSV export for transaction history (#337)
A new screen has been added in the iOS and the android app to export the 
payments history as a CSV file. Both apps delegate the csv serialization to a 
custom `CsvWriter` utility file in the shared module. Writing the file to disk is
handled by the front apps.

The payments exported can be filtered by date and the data are only a subset
of all the payment information contained in the database. Included in the CSV 
are the completion date, the amount and the fee (btc and fiat), some generic 
context about the payments, and the description of the payments.

Only successfully completed payments are exported.

Co-authored-by: Dominique <5765435+dpad85@users.noreply.github.com>