Process lnid FCM notifications on Android
The FCM service now handles the `invoice` notifications
sent by the lnid service and sends back the invoice to a
hard-coded placeholder service. This is a work in progress.

Registered new notification channels and added an helper to
start notifications easily.

The FCM service now checks FCM message priority to avoid
starting a foreground service if the priority is not high.
(new constraints from Android 13).

The FCM library has been upgraded.