Implement database support for splices
/!\ This change is a work-in-progress based on a fleeting lightning-kmp
version. The frontend apps are not updated.

OutgoingPayment is now a sealed class, which can be a lightning payment
(that we already support, and store in the `outgoing_payments` tables)
or a splice-out. Naming has been updated.

A new table `splice_outgoing_payments` has been added to store splice-outs.

A new incoming payment part `SpliceIn` has been added, also stored as a
json like the other received-with parts.

WalletPaymentId has a new type for splice-outs.

Several methods listing payments have been removed from the payments db as
they were not used.

`createdAt` and `completedAt` are now interfaced in WalletPayment.

The method calculating the channel's balance now relies on a method defined
in lightning-kmp.