(android) Add wallet info and liquidity policy screens
The wallet info screens display the swap-in and final wallets balance
and xpub, as well as their respective unconfirmed utxos in details
screens. This lets users tracks what happens in their on-chain wallets.

The liquidity policy screens lets users set what policy to use and what
fee to pay when receiving payments. This screen is a WIP and may change.