Update database and keymanager following changes in lightning-kmp
The key manager now exposes a straightforward API, especially for
the swap-in/final wallets data.

OutgoingPayment has a new ChannelCloseOutgoingPayment type to store
channels closing. The ClosingTxPart object has been removed.

We also now use a confirmed_at timestamp to track if an on-chain
payment is confirmed.

A new table has been added to links transactions with incoming or
outgoing payments.

This commit is a WIP.