(android) Streamline android's network monitor implementation
The coroutine monitoring connection is launched in the init. The
flag tracking the state of the monitoring have been removed as
they are prone to race conditions and we don't (yet) need to disable
or stop the monitor on Android.

Some debugging has also been added in the home screen connection
dialog, to track the state of the network.