Update database following locked/confirmed changes
On-chain payments now have two timestamps, locked_at and
confirmed_at. First one is when the funds are locked, i.e.
are deemed safe to use by lightning-kmp according to its
internal rules. Confirmed_at is the timestamp when the
underlying transaction has been confirmed.

Also, expectedAmount has been removed from IncomingPayments.
Pay-to-open are rejected when channels are not ready.