Add system to persist notifications
This commits adds a new notifications table that is used to
store all kinds of notifications. For now we only store
notifications related to payments that have been rejected.

This commmit also updates to the latest kmp version, where
the liqudity policy of the node params can be updated. As such
a new method has been added in Peer to update the liquidity
policy on the fly.