Add flag in PeerManager when upgrading is required
The peer may send an `UpgradeRequired` event when an old Phoenix
connects and request channels that use a type that this version
of Phoenix cannot handle (e.g., Phoenix does not have the adequate
features).

In that case the app should show a blocking message that signals
that the app should be updated. Note that there may NOT yet be a
version of Phoenix that can handle these new channels. For example
if the user switches from iOS to Android, and the Android app is
behind iOS and not ready yet.