(legacy) Use multisig swap for migration and stagger closings
A new version of eclair-core is needed to obtain the swap-in
multisig address.

Migration is now done channel by channel to prevent herd effect.
Also we monitor that the closing transactions are published
before completing the migration process. This is important to
have one single swap.

The UI of the migration dialog has also been improved.