Support multisig swap wallet logic from lightning-kmp
The balance manager has been simplified. We now use a helper
method from lightning-kmp to filter utxos reserved for channels.

The flows for the swap/final wallets have been moved to the
PeerManager.

The wallet balance object now distinguishes between weakly and
deeply confirmed utxos. This is useful for the swap-in wallet
where weakly confirmed utxos cannot be used for swaps.