(android) Split string resources into two files
This should help the translation effort by making it obvious
which resources should be translated first. The important_strings
file contains mostly contextual information and instructions that
guide the user through key parts of the app. This file should be
kept as small as possible.

Resources found in the default `strings.xml` are less important
and may be skipped or translated later on.