Add migration process for iOS wallets using legacy channels (#395)
An helper class has been added in the shared module. When the
app starts with at least 1 active channel which is not dual funding, 
a migration process is started.

This migration process is in essence similar to the legacy android 
app migration. Existing channels are closed to the swap-in potentiam
address. This enables the iOS app to use dual funded channels and
splicing.

Migration is mandatory since the app is not able to use legacy 
channels. The iOS app displays a screen that guides the user 
through migration.

Migration is completed once the closing txs are visible in the
swap-in wallet.

Note that dust channels do not return an error, instead a warning 
will be displayed (this will be done in another PR).

---------

Co-authored-by: Robbie Hanson <304604+robbiehanson@users.noreply.github.com>